+++ packages/werkbank/lib/src/addon_api/src/addon_control_sections/tabs/inspect_control_section.dart
@@ -13,7 +13,7 @@ class InspectControlSection extends AddonControlSection {
 }
 
 class InspectControlSectionAccessor extends AddonControlSectionAccessor
-    with UseCaseAccessorMixin {
+    with UseCaseAccessor {
   const InspectControlSectionAccessor();
 
   @override