+++ packages/werkbank/lib/src/addon_api/src/addon_layer_entries/layers/use_case_fitted_layer_entry.dart
@@ -16,7 +16,7 @@ class UseCaseFittedLayerEntry extends AddonLayerEntry {
 }
 
 class UseCaseFittedLayerEntryAccessor extends AddonLayerAccessor
-    with MaybeWerkbankAppAccessor, UseCaseAccessorMixin {
+    with MaybeWerkbankAppAccessor, UseCaseAccessor {
   const UseCaseFittedLayerEntryAccessor();
 
   @override